Implement vdi_revert and expose VDI_REVERT for EXT and NFS
This is a bit hacky.

In FileSR we extend _do_snapshot with a new parameter 'dest_uuid'
which defaults to None. If dest_uuid is not None then this is
used as the name of the file created in the SNAPSHOT_DOUBLE case.

We assume (hacky hacky) that if SNAPSHOT_DOUBLE and dest_uuid then
we should not introduce the new VDI, since this is the revert case
and the VDI already exists. We wish to keep the VDI record stable
in Xapi and avoid deleting and re-creating it.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>